Annotate fallthrough and promote two declaration warnings
Clang 22 reported 33 case labels reached by falling through, because it
does not accept fallthrough comments. Intended fallthroughs now carry
[[fallthrough]], and cases that only fell into a break-only label end in
break. Two were bugs: SKILL_DIRT_KICK fell into the berserker rage check,
so a level 20, dexterity 17 character without 15 rogue levels passed it
as a berserker, and the object editor's special ability value prompts
fell through every later prompt into the SYSERR default case.

-Wredundant-decls and -Wnested-externs reached zero on GCC 16.2 and
Clang 22.1.8 in the previous change and build without warnings at the
baseline tier on GCC 13 and Clang 18, so they move from the migration
tier to the baseline, where the strict jobs make them errors.

+++ src/ai_service.c
@@ -926,7 +926,7 @@ static char *make_api_request_single(const char *prompt)
   else
   {
     log("AI Service: CURL error: %s", curl_easy_strerror(res));
-    AI_DEBUG("CURL error details: code=%d, message=%s", res, curl_easy_strerror(res));
+    AI_DEBUG("CURL error details: code=%d, message=%s", (int)res, curl_easy_strerror(res));
   }
 
   /* Cleanup */
